Change trip authorization
Updated rules:

- [NEW] Visitor is redirected to sign in if there is no token.
- Then if there is no trip, of course it's a 404
- Then if:
  1. user is signed in AND is not organizer OR has an invalid token, it's a 404
  OR
  2. visitor has an invalid token, it's a 404

Plus, extract trip authorization into tis own module.
